Accept floats for existing dividends and employment income
Type annotations changed from int to float for existing_income and
existing_dividends parameters across CLI prompts, calculators, and
core tax functions. Added prompt_float() to CLI. Updated format_gbp
to handle float amounts with rounding.

Added test coverage with float input values for:
- calc_income_tax with existing_income=30000.50
- calc_income_tax with existing_income=50000.75 (band boundary)
- calc_dividend_tax with existing_income=30000.50
- InsideIR35Calculator with existing_income=30000.50
- InsideIR35Calculator with existing_dividends=15000.75
- format_gbp with float values

+++ payday/calculators/inside_ir35.py
@@ -21,8 +21,8 @@ class InsideIR35Calculator:
         working_days: int,
         umbrella_margin_weekly: int = 25,
         start_month: int | None = None,
-        existing_income: int = 0,
-        existing_dividends: int = 0,
+        existing_income: float = 0,
+        existing_dividends: float = 0,
         salary_sacrifice: int = 0,
     ) -> SalaryBreakdown:
         """Inside IR35: Assignment → Er costs → gross → IT + EE NI + Pension → 20-day.
